feat: add Docker setup
This adds new feature to allow the bot to be run via Docker. This is
useful for situations where you want to:

* control the lifecycle of the bot with Docker daemon instead of system
  cron daemon
* abstract away the environment

The README was updated with this new information and two new config
flags were added to facilitate this setup.

The old way of launching bot is still applicable.